fix(migrate): always use raw perspective for migrations
perspective=raw is currently default for all current API versions, but that might not be true in for all future versions. For migrations, perspective=raw is the most sensible default, so this PR explicitly sets perspective=raw when fetching documents for migration. If there's a compelling use case for having user-provided perspectives, we can always consider adding support for that that later, but this change ensures that perspective=raw is always used, no matter what the default is backend side.