feat(core): add default bold/italic markdown shortcuts to PTE
The `MarkdownPlugin` now ships with automatic bold/italic shortcuts that you
can opt in to by defining an `italicDecorator` and/or a `boldDecorator`. The
italic decorator is then mapped to `*`/`_` pairs and the bold decorator is
mapped to `**`/`__` pairs.

Following the standard behaviour that is observed in most other text editors,
the mapping only occurs when the end char is inserted. For example, typing
`*foo` and then `*` bar makes `foo` italic, but typing `foo*` and then adding a
`*` before `foo*` does nothing.

After a mapping occurs, if you didn't like it, you can undo it as expected. And
if you don't move your cursor, you can even undo it by hitting Backspace.