fix(core): flag `application/x-portable-text` as portable text
This makes it possible to drag inline objects in the editor without getting a
"Can't upload any of these files here" error.

---

Drag and drop in PTE is still very lackluster:

1. Block objects like images can be dragged around, but this is purely
   implemented by maintaining an internal drag state, and dragging these
   objects doesn't use the `application/x-portable-text` MIME type yet.
2. Inline objects can be dragged more freely. They are stored using
   `application/x-portable-text` and even support being dragged from one editor
   to another.

`application/x-portable-text` is what PTE uses when serializing and
deserializing to and from the clipboard. And it's also the MIME type used for
drag and drop going forward. And soon block objects will be stored on this MIME
type as well.