feat(core): add default `code` and `strike-through` markdown shortcuts to PTE
The `MarkdownPlugin` now also handles the conversion of "\`" pairs as well as
the conversion of "~~" pairs. All you need to do is tell it what decorator to
use for the conversion. In our case, we go with the default decorator names so
anyone who has these configured get the shortcuts.