fix(core): make PTE active by default
There's been some feedback internally and externally that it would be nice if
the Portable Text Input was active by default. So I think we should try it.

Therefore, the default value of the `PortableTextInput.initialActive` prop has
been flipped from `false` to `true`.

The only downside is that the editor might scrolljack you on a small
device/viewport, but this does not appear to be a good reason to make the
editor inactive by default for everyone, especially not when most users are on
desktop devices.