feat(core): allow configuring PTE plugins
This commit adds a `renderPlugins` prop on the `PortableTextInput` that can be
used to inject Portable Text Editor Plugins in the editor.

PTE Plugins are just React components. And the reason for this is that for
complex Plugins, you'll want to maintain some local state and the most
idiomatic way to do that is using React. The `DecoratorShortcutPlugin` is an
example of such plugin (it contains an entire XState state machine). Another
reason is that, down the line, we want to render UI elements through Plugins
(think: Emoji Picker).

It should be noted that rendering UI elements in PTE Plugins hasn't been fully
thought through. For example, we might want to expose some affordance to
control *where* the UI elements are rendered. I think, we can address this
later as the `renderPlugins` is marked as `@beta` and isn't intended to be
publically announced yet.

The most important thing right now is that you have an entry point to
get the `Editor` instance using `useEditor()` and add Behaviors using
`editor.registerBehavior(...)`.