fix(core): make PTE inactive by default on mobile
This way you can easily scroll a document on a mobile phone without a long
Portable Text Input hijacking the scroll.

Should be noted that this check isn't bulletproof since the user might be on an
iPad or other large touch device where scroll-jacking wouldn't be an issue, but
it feels like a good enough heuristic.