fix(core): restore intended func. for 'directUploads' config param
The config.form.image.directUploads and similar for 'file' was introduced when we created assetSources in v2.
The reason for this, is to prohibit local uploads and force users to only use assets from an existing DAM.
Somewhere along the v3 journey this option was misinterpreted and also disabled selecting from
the asset sources and using an external asset in the input (so treated simlilar to read only).

This breaks the intention of the config parameter, and I have restored it to how it should work.

I have also added a reducer function for the config parameters so they can be properly set through
plugins.