fix(core): remove `renderEditable` from PTE inputs again
`renderEditable` was added as a `@hidden` `@beta` feature in #6627 in order to
provide Canvas (then Create) a way to configure PTE at the lowest level.
However, Canvas has been using Standalone PTE for more than half a year now and
as such, there is no need for this API anymore. In fact, the `renderEditable`
API poses a problem to the Standalone PTE itself since it exposes the entire
prop surface of the `<PortableTextEditable />` offered by this library. The
more Standalone PTE APIs we expose directly through Studio, the harder it is
going to be to move `@portabletext/editor` forward since breaking these APIs,
in most cases, will be impossible (since that would introduce breaking changes
in Studio).

On top of that, most of the use cases Canvas had for `renderEditable` (for
example hooking into text insertions and altering their outcome) have been
replaced with the Behavior API, which, in the near future, is going to be the
recommended way to extend PTE through Studio as well.

To be on the safe side and to shield Standalone PTE from too much direct
exposure through Studio, this commit removes `renderEditable` again.