feat(schema): place builtin types in a separate registry
This expands the `Schema` type to have a concept of "parent registry".
We're using this to define the built-in types in a single registry and
then in `createSchema` we rather refer to this existing registry. This
avoid us having to construct duplicate schema types for different
schemas/workspaces.

It also makes it easier to reason about which types were actually
defined (`getLocalTypeNames`) whereas which were inherited.