feat(sanity): allow new, unpersisted, documents to move between perspectives seamlessly
Previously, if a user switched perspective after creating a new
document—but that document wass never actually created in the dataset—they
would have to manually add it to the newly selected perspective in order
to continue working on it.

This occured when a user created a document but didn't edit it.

This workflow is quite clunky; the user must explicitly move the
document around, even though it doesn't represent a real document in
the dataset.

This commit allows the user to simply switch perspectives and begin
editing the new document, as long as the document has not yet been
created in the dataset.