fix(core): use `data-list-index` on PTE blocks to improve list counts
Calculating the correct list index using CSS is extremely tricky. Therefore,
PTE now ships a `data-list-index` on each list block (a block with both
`listItem` and `level`) by default. (Notice how the blocks now also have
`data-level` as well.)

While a "list index" isn't a part of the stored Portable Text nor the protocol
itself, it's universally understood that a block with `listItem` and `level`
*is* a list item. At the same time, we want to make it as seamless as possible
to adopt and style the Portable Text Editor, and thus, `data-list-index` came
into play.

Under the hood, PTE does an efficient calculation of where a list block is
positioned in its current list. This not only simplifies the CSS needed to keep
track of the CSS counters. It also fixes tricky edge cases like this one:

```
1. foo
      1. bar
    i. baz
      1. new list
```

Before this change, "new list" would have count 2, but because of the indented
parent, the correct count is 1 (Word and Google Docs agree).

Only if the parent is outdented, should the count be 2:

```
1. foo
      1. bar
        a. baz
      2. continuing "bar" list
```