feat(sanity): allow `groq2024` search strategy to match on referenced `_id`
This change allows the `groq2024` search strategy to surface documents
that contain a reference to a document whose `_id` exactly matches
the search query.

For example: searching for "some-exact-id" will yield the document with
that `_id` *along with* all documents that contain a `_ref`
(at any depth) set to that id.

Note that the entire search query must match the id. We do not perform
any tokenisation at the moment.