fix(sanity): do not attempt to provision media library
Users are no longer permitted to provision a media library themselves.
Therefore, Studio must not attempt to provision a media library for
users on the fly.

Instead, a warning is shown when attempting to interact with a
Media Library asset source should no media library exist for the
current organisation.