fix(sanity): explicitly unset legacy inspect parameter (#10208)
### Description

Small change to fix an issue we've been working around in Presentation which I think makes sense to fix upstream.

When closing the legacy inspector panel, instead of using `omit` to remove the `inspect` parameter from the params object, we now explicitly set it to `undefined`. This just means that Presentation doesn't have to check if the parameter is present each time its pane router receives new params.

### What to review

- Check the change in `useDocumentPaneInspector.ts` where `inspect: undefined` is set instead of omitting the property is correct.

### Testing

The inspector functionality should be unchanged, ensure it opens and closes as before.

### Notes for release

N/A