feat(sanity): add UI control for toggling visibility of inline changes
This change makes inline diffs available for all workspaces, regardless
of whether Advanced Version Control is switched on.

- In the main document editor, users can now toggle the visibility
of inline diffs as they please. The toggle is available as an action in
the document editor context menu.
- In the side-by-side document comparison view, inline diffs are always
visible.

A new input prop named `displayInlineChanges` has been added to indicate
whether input components should display an inline diff.

The state is stored as a router sticky param, so it persists when
navigating (e.g. between documents).