refactor(sanity): require complete `SanityDocument` for `compareValue`
This provides better compatibility with other parts of the code that we
may wish to pass `compareValue` to, which are typed as
`FormDocumentValue`. `Partial<SanityDocument>` doesn't require the
`_type` and `_id` fields, whereas `FormDocumentValue` does.

We know these values are present in the data, because `compareValue`
reflects an existent document. The type now reflects that.