feat(core): add configurable `typography` plugin to PTE inputs
The `@portabletext/plugin-typography` plugin has been added as a built-in
plugin for PTE, configurable using the `portableText.plugins.typography` prop.

https://www.npmjs.com/package/@portabletext/plugin-typography

The default configuration adds a `guard` to the plugin which prevents it from
running inside `'code'` decorators. The plugin also has available props to
configure a `preset` or configure `enabled` and `disabled` rules. All of
these props, along with the `guard`, are exposed.

The default `preset` is `'default'` and enables the default set of rules: https://www.npmjs.com/package/@portabletext/plugin-typography#rules-included

Note: In order to allow configuring a second plugin, the existing `markdown`
has been made optional and it's now required to spread `...props.plugins` in
order to not accidentally overwrite any built-in plugin.