fix(core): flush pending Portable Text Input changes on unmount
For performance reasons, PTE `'mutation'` events are debounced. This means that
it will take a few ms from after you've stopped typing until the changes are
emitted and can be stored in Content Lake.

However, this could cause a problem if you navigate away quickly after typing,
and this scenario isn't even that unlikely if the user is typing in a dialog
that is then quickly closed. This can result in lost data as well a lack of
trust in the overall system.

To mitigate this, latest version of PTE, `v3.0.6`, brings a patch to flush
pending `'patch'` and `'mutation'` events when the editor detects that it's
being unmounted.

This commit upgrades PTE to that version.